[Refactor] 모든 페이지에서 쓰이는 ProductItem컴포넌트 별도의 파일로 관리하기 위한 리팩토링
