+++ src/page/Main.js
@@ -4,6 +4,7 @@ import Header from "../component/Header";
 import Footer from "../component/Footer";
 import { IoClose } from "react-icons/io5";
 import { FaStar } from "react-icons/fa";
+import ProductItem from "../component/ProductItem";
 
 const Container = styled.div`
   display: flex;
@@ -119,59 +120,6 @@ const ProductModal = styled.div`
     }
   }
 `;
-
-// 이미지를 보여주는 컴포넌트 생성
-// item은 클릭은 그 상품자체의 정보를 담고있음
-// onClick은 showModal과 selectedImage 의 상태를 바꾸어주는 clickModal 핸들러 함수를 전달받음
-// handleBookmarkClick는 현재 클릭한 사진의 bookmark 이미지 클릭시 isBookmarked의 값을 바꿔주는 핸들러함수
-const ProductItem = ({ item, onClick, handleBookmarkClick }) => {
-  return (
-    <div className="itemContainer" onClick={onClick}>
-      <FaStar
-        onClick={(event) => {
-          event.stopPropagation(); // .itemContainer의 onClick 핸들러함수의 이벤트 캡쳐링을 방지하고자 썻으나, 없어도 정상작동(리팩토링 전에는 필요했음)
-          handleBookmarkClick(item);
-        }}
-        fill={item.isBookmarked ? "#FFD361" : "white"} // isBookmarked 여부에 따라 별 색상 변경
-      />
-      <img src={item.image_url} alt={item.title} />
-      <div className="descriptionContainer">
-        <div className="subDescriptionContainer">
-          <div className="descriptionItem">
-            {item.type === "Product"
-              ? item.title
-              : item.type === "Category"
-              ? `# ${item.title}`
-              : item.type === "Exhibition"
-              ? item.title
-              : item.type === "Brand"
-              ? item.brand_name
-              : ""}
-          </div>
-          <div className="descriptionItem">
-            {item.type === "Exhibition" ? item.sub_title : ""}
-          </div>
-        </div>
-        <div className="subDescriptionContainer2">
-          <div className="descriptionItem">
-            {item.type === "Product"
-              ? `${Number(item.price).toLocaleString()}원`
-              : item.type === "Brand"
-              ? "관심고객수"
-              : ""}
-          </div>
-          <div className="descriptionItem">
-            {item.type === "Product"
-              ? `${item.discountPercentage}%`
-              : item.type === "Brand"
-              ? item.follower
-              : ""}
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
 
 const Main = ({
   productData,
@@ -291,7 +239,7 @@ const Main = ({
           </ProductModal>
         )}
 
-        {/*북마스 리스트 랜더링구간 */}
+        {/*북마크 리스트 랜더링구간 */}
         <ItemWrapper>
           <div className="title">북마크 리스트</div>
           <div>
